+++ Sourcery/Models/Variable.swift
@@ -36,7 +36,7 @@ final class Variable: NSObject, SourceryModel, Typed, Annotated {
     var attributes: [String: Attribute]
 
     /// Underlying parser data, never to be used by anything else
-    // sourcery: skipEquality, skipDescription, skipCoding
+    // sourcery: skipEquality, skipDescription, skipCoding, skipJSExport
     internal var __parserData: Any?
 
     init(name: String = "",