+++ SourceryTests/Stub/Stubs.swift
@@ -15,6 +15,7 @@ private class Reference {}
 enum Stubs {
     private static let basePath = Bundle(for: Reference.self).resourcePath.flatMap { Path($0) }!
     static let swiftTemplates = basePath + Path("SwiftTemplates/")
+    static let jsTemplates = basePath + Path("JavaScriptTemplates/")
     static let sourceDirectory = basePath + Path("Source/")
     static let sourceForPerformance = basePath + Path("Performance-Code/")
     static let resultDirectory = basePath + Path("Result/")